Refactor ServiceForm for advanced auth flow
