+++ airbyte-webapp/src/views/Connector/ServiceForm/components/Sections/LegacyAuthSection.tsx
@@ -1,0 +1,17 @@
+import React from "react";
+
+import { FeatureItem, WithFeature } from "hooks/services/Feature";
+import { SectionContainer } from "./common";
+import { AuthButton } from "../AuthButton";
+
+export const LegacyAuthSection: React.FC = () => {
+  return (
+    <WithFeature featureId={FeatureItem.AllowOAuthConnector}>
+      {
+        <SectionContainer>
+          <AuthButton />
+        </SectionContainer>
+      }
+    </WithFeature>
+  );
+};
