+++ airbyte-webapp/src/views/Connector/ServiceForm/components/Sections/PropertySection.tsx
@@ -4,9 +4,9 @@ import { useField } from "formik";
 
 import { LabeledToggle, TextWithHTML } from "components";
 import { FormBaseItem } from "core/form/types";
-import { Label } from "./Property/Label";
-import { Control } from "./Property/Control";
-import { useServiceForm } from "../serviceFormContext";
+import { Label } from "../Property/Label";
+import { Control } from "../Property/Control";
+import { useServiceForm } from "../../serviceFormContext";
 
 const PropertySection: React.FC<{ property: FormBaseItem; path?: string }> = ({
   property,
