+++ airbyte-webapp/src/views/Connector/ServiceForm/components/Sections/common.tsx
@@ -1,0 +1,5 @@
+import styled from "styled-components";
+
+export const SectionContainer = styled.div`
+  margin-bottom: 27px;
+`;
