+++ airbyte-webapp/src/views/Connector/ServiceForm/components/utils.ts
@@ -1,0 +1,18 @@
+import { FormBlock } from "core/form/types";
+import { naturalComparator } from "utils/objects";
+
+export function OrderComparator(a: FormBlock, b: FormBlock): number {
+  const aIsNumber = Number.isInteger(a.order);
+  const bIsNumber = Number.isInteger(b.order);
+
+  switch (true) {
+    case aIsNumber && bIsNumber:
+      return (a.order as number) - (b.order as number);
+    case aIsNumber && !bIsNumber:
+      return -1;
+    case bIsNumber && !aIsNumber:
+      return 1;
+    default:
+      return naturalComparator(a.fieldKey, b.fieldKey);
+  }
+}
