+++ airbyte-webapp/src/views/Connector/ServiceForm/useBuildForm.tsx
@@ -1,7 +1,7 @@
 import { AnySchema } from "yup";
 import { useCallback, useEffect, useMemo, useState } from "react";
 import { useFormikContext } from "formik";
-import { JSONSchema7 } from "json-schema";
+import { JSONSchema7, JSONSchema7Definition } from "json-schema";
 import flatten from "flat";
 import merge from "lodash.merge";
 
@@ -14,38 +14,55 @@ import { ConnectorDefinitionSpecification } from "core/domain/connector";
 import { FeatureItem, useFeatureService } from "hooks/services/Feature";
 import { applyFuncAt, removeNestedPaths } from "core/jsonSchema";
 
+function upgradeSchemaLegacyAuth(
+  connectorSpecification: Required<
+    Pick<
+      ConnectorDefinitionSpecification,
+      "authSpecification" | "connectionSpecification"
+    >
+  >
+) {
+  const spec = connectorSpecification.authSpecification.oauth2Specification;
+  return applyFuncAt(
+    connectorSpecification.connectionSpecification,
+    spec.rootObject ?? [],
+    (schema) => {
+      // Very hacky way to allow placing button within section
+      (schema as any).is_auth = true;
+      const schemaWithoutPaths = removeNestedPaths(
+        schema,
+        spec.oauthFlowInitParameters ?? [],
+        false
+      );
+
+      const schemaWithoutOutputPats = removeNestedPaths(
+        schemaWithoutPaths,
+        spec.oauthFlowOutputParameters ?? [],
+        false
+      );
+
+      return schemaWithoutOutputPats;
+    }
+  );
+}
+
 function useBuildInitialSchema(
   connectorSpecification?: ConnectorDefinitionSpecification
-): JSONSchema7 | undefined {
+): JSONSchema7Definition | undefined {
   const { hasFeature } = useFeatureService();
 
   return useMemo(() => {
-    if (
-      hasFeature(FeatureItem.AllowOAuthConnector) &&
-      connectorSpecification?.authSpecification
-    ) {
-      const spec = connectorSpecification.authSpecification.oauth2Specification;
-      return applyFuncAt(
-        connectorSpecification.connectionSpecification,
-        spec.rootObject ?? [],
-        (schema) => {
-          // Very hacky way to allow placing button within section
-          (schema as any).is_auth = true;
-          const schemaWithoutPaths = removeNestedPaths(
-            schema,
-            spec.oauthFlowInitParameters ?? [],
-            false
-          );
-
-          const schemaWithoutOutputPats = removeNestedPaths(
-            schemaWithoutPaths,
-            spec.oauthFlowOutputParameters ?? [],
-            false
-          );
-
-          return schemaWithoutOutputPats;
-        }
-      );
+    if (hasFeature(FeatureItem.AllowOAuthConnector)) {
+      if (
+        connectorSpecification?.authSpecification &&
+        !connectorSpecification?.advancedAuth
+      ) {
+        return upgradeSchemaLegacyAuth({
+          connectionSpecification:
+            connectorSpecification.connectionSpecification,
+          authSpecification: connectorSpecification.authSpecification,
+        });
+      }
     }
 
     return connectorSpecification?.connectionSpecification;
@@ -80,7 +97,7 @@ function useBuildForm(
   };
 }
 
-const useBuildUiWidgets = (
+const useBuildUiWidgetsContext = (
   formFields: FormBlock[] | FormBlock,
   formValues: ServiceFormValues,
   uiOverrides?: WidgetConfigMap
@@ -165,7 +182,7 @@ const usePatchFormik = (): void => {
 export {
   useBuildForm,
   useBuildInitialSchema,
-  useBuildUiWidgets,
+  useBuildUiWidgetsContext,
   useConstructValidationSchema,
   usePatchFormik,
 };
